Moved Blazor startup code to new service
